Use a destroy function to free all the resources.
The handle returned by dlopen() is now saved and the library is unloaded
on exit.
We also load symbol "dynamic_driver_destroy" which is a function with no
arguments returning void. It is called on exit.
The load_library() is now split into two functions, one function is
specifically used to load symbols.