Pass all necessary pointers in a packed structure.
Access to the dns_dyndb_arguments_t structure is done through a set of
dns_dyndb_set_* and dns_dyndb_get_* functions. This will make it
unnecessary to break the ABI if someone designing a plug-in will have a
need for something more than there currently is. All that will be needed
in that case will be an addition of set/get functions.